+++ app/wyzebridge/wyze_stream.py
@@ -1,9 +1,11 @@
 import contextlib
 import json
 import multiprocessing as mp
+import zoneinfo
 from collections import namedtuple
 from ctypes import c_int
 from dataclasses import dataclass
+from datetime import datetime
 from enum import IntEnum
 from queue import Empty, Full
 from subprocess import PIPE, Popen
@@ -15,7 +17,7 @@ from wyzebridge.bridge_utils import env_bool, env_cam
 from wyzebridge.config import BRIDGE_IP, COOLDOWN, MQTT_TOPIC
 from wyzebridge.ffmpeg import get_ffmpeg_cmd
 from wyzebridge.logging import logger
-from wyzebridge.mqtt import send_mqtt, update_mqtt_state, wyze_discovery
+from wyzebridge.mqtt import publish_discovery, send_mqtt, update_mqtt_state
 from wyzebridge.webhooks import ifttt_webhook
 from wyzebridge.wyze_api import WyzeApi
 from wyzebridge.wyze_commands import GET_CMDS, PARAMS, SET_CMDS
@@ -100,7 +102,7 @@ class WyzeStream:
             logger.error(f"{self.camera.nickname} may not support multiple streams!!")
             # self.state = StreamStatus.DISABLED
         self.options.update_quality(self.camera.is_2k)
-        wyze_discovery(self.camera, self.uri)
+        publish_discovery(self.uri, self.camera)
 
     @property
     def state(self):
@@ -234,7 +236,7 @@ class WyzeStream:
             self.update_cam_info()
         if self.camera.camera_info and "boa_info" in self.camera.camera_info:
             data["boa_url"] = f"http://{self.camera.ip}/cgi-bin/hello.cgi?name=/"
-        return data | self.camera.dict(exclude={"p2p_id", "enr", "parent_enr"})
+        return data | self.camera.model_dump(exclude={"p2p_id", "enr", "parent_enr"})
 
     def update_cam_info(self) -> None:
         if not self.connected:
@@ -271,6 +273,18 @@ class WyzeStream:
                 return self.api.get_pid_info(self.camera, "P3")
             run_cmd = payload if payload == "restart" else f"{cmd}_{payload}"
             return dict(self.api.run_action(self.camera, run_cmd), value=payload)
+        if cmd == "time_zone" and payload and isinstance(payload, str):
+            try:
+                zone = zoneinfo.ZoneInfo(payload)
+            except zoneinfo.ZoneInfoNotFoundError:
+                return {"response": "invalid time zone"}
+            if offset := datetime.now(zone).utcoffset():
+                return dict(
+                    self.api.set_device_info(
+                        self.camera, {"device_timezone_city": zone.key}
+                    ),
+                    value=int(offset.total_seconds() / 3600),
+                )
 
         if self.state < StreamStatus.STOPPED:
             return {"response": self.status()}